==> analysis/00 Actualizar Features PP.ipynb <==
import numpy as np
import pandas as pd
import csv
import os

## Actualizar OD

os.chdir('./../00 Insumos/05 Object Detection')
os.listdir()

with open('od_class_dict.csv') as file:
    reader = csv.reader(file)
    next(reader) 
    class_dict = {int(rows[0]):rows[1] for rows in reader}
#     class_dict.pop('key')
    
od = pd.read_csv('output_object_detection_stgo.csv')


od['class'] = np.vectorize(class_dict.get)(od['class'])
od['aux_val'] = 1
# od.head()


## Agrupar: Contar cada objeto por imagen
Pivot = pd.DataFrame(pd.pivot_table(od,
                                    index=['img_id'], 
                                    columns=['class'], 
                                    values=['aux_val'], 
                                    aggfunc='count').sum()).sort_values([0],ascending=False)
Pivot = Pivot.reset_index().drop(columns=['level_0'])

## Seleccionar top 15 objetos
top15_objects = Pivot['class'].head(20).tolist()
# Pivot.head(15)
# for i in Pivot['class'].values:
#     print(i)

## 
od_filtered = od[(od['class'].isin(top15_objects))]
od_features = pd.pivot_table(od_filtered,
               columns=['class'],
               index=['img_id'],
               values=['aux_val'],
              aggfunc='sum').fillna(0)
od_features.columns = [ 'od_' + i for i in od_features.columns.get_level_values(1) ]
od_features.index = pd.Index(od_features.index, name='ID')

od_features

# os.chdir('./../02 Stgo')
# os.listdir()
# od_features.to_csv('OD.csv')

## TENDENCIA DE LA DETECCION DEOBJETOS
# Pivot2 = pd.DataFrame(pd.pivot_table(od,
#                                     index=['img_id'], 
#                                     columns=['class'], 
#                                     values=['aux_val'], 
#                                     aggfunc='count'))

## Tabla tendencias solo con valores de imagenes con almenos 1 objeto
# Pivot2.describe().T.sort_values(['count'],ascending=False).iloc[:15,0]

## Tabla tendencias sobre todas las imagenes observadas
# Pivot2.fillna(0).describe().T.sort_values(['mean'],ascending=False).iloc[:15,:]

# Tabla conjunta
# pd.concat([round(Pivot2.describe().T.sort_values(['count'],ascending=False).iloc[:15,0],2),
#            round(Pivot2.fillna(0).describe().T.sort_values(['mean'],ascending=False).iloc[:15,1:],2)], axis=1).sort_values(['count'],ascending=False)
# .to_latex()

## Actualizar Features

os.chdir('./../00 Insumos/01 PP')
os.listdir()


HLF = pd.read_csv('HLF.csv', index_col='ID')
LLF = pd.read_csv('LLF.csv',index_col='ID')
nans = pd.read_csv('nans.csv',index_col='ID')
od = pd.read_csv('OD.csv',index_col='ID')

features = pd.concat([nans, LLF, HLF, od], axis=1).reset_index()
features = features.rename(columns ={'index':'ID'})

features.to_csv('all_features_imagenes.csv')

features
# .columns

od

## Numer of objects per class

features.to_csv('new_features_imagenes.csv',index=False)
