==> analysis/01 Actualizar respuestas Wekun.ipynb <==
# Analisis Base de datos

import os
import pandas as pd
import numpy as np

## Open saved DB

os.chdir('./../data/04 Wekun')
users = pd.read_csv('users_16-09-2019.csv')
answers = pd.read_csv('answers_16-09-2019.csv')

users.columns

answers['choice'] = answers['choice'].replace({'image_1':1,
                                   'image_2':2,
                                   '=':3,
                                   '= ':3,
                                   'X':99})
answers['choice'] = answers['choice'].astype(int)

users.rename(columns={'id':'user'}, inplace=True)

       
users['gender'] = users['gender'].replace({'M':0,
                                   'F':1,
                                  'O':99})
users['education'] =  users['education'].replace({'es1':1,'en1':1,
                                          'es2i':1,'en2i':1,
                                          'es3i':2,'en3i':2,
                                          'es4i':3,'en4i':3,
                                          'es5i':3,'en5i':3,
                                          'es6i':5,'en6i':5,
                                          'es2':2,'en2':2,
                                          'es3':3,'en3':3,
                                          'es4':4,'en4':4,
                                          'es5':5,'en5':5,
                                          'es6':6,'en6':6})
users = users.astype({'user_id':str,
              'user':int,
              'age':float,
              'gender':int,
              'education':int,
              'transport':int})
age_outlayers = users[(users["age"]>= 100) | \
                     (users["age"]< 0)]

users = users[(users["age"]< 100) & \
         (users["age"]>= 0)]

users = users.astype({'age':int})

df = pd.merge(answers,users, how='inner', on='user_id', sort=False, indicator=False,suffixes=('_ans', '_usr'))

df.head()

df.shape

## Place Pulse

# PP = pd.read_csv('./../05 Benchmark/PP2_votes.csv')

# PP['winner'] = PP['winner'].replace({'left':1,
#                                     'right':2,
#                                     'equal':3})

# PP.head()

# models_db = PP[['left_id', 'right_id','winner', 'category']]
# models_db.columns = pd.Index(['img_1', 'img_2', 'choice','question'])

# Generar archivo .DAT

os.chdir('./../01 PP')

features = pd.read_csv('all_features_imagenes.csv')
models_db = df

models_db = models_db.merge(features, how='inner', left_on='img_1', right_on='ID', sort=False, suffixes=('', '_izq'))
models_db = models_db.merge(features, how='inner', left_on='img_2', right_on='ID', sort=False, suffixes=('_izq', '_der'))
models_db = models_db.drop(columns=['img_1','img_2','ID_izq','ID_der'])

models_db.shape

outliers = models_db[(models_db['NA_izq'] != 0) | (models_db['NA_der'] != 0)]
#                      | \
#              (models_db['outlier_bn_izq'] != 0) | (models_db['outlier_bn_der'] != 0) | \
#              (models_db['outlier_hs_izq'] != 0) | (models_db['outlier_hs_der'] != 0) | \
#              (models_db['outlier_hlf_izq'] > 1) | (models_db['outlier_hlf_der'] > 1)]

models_db = models_db[(models_db['NA_izq'] == 0) & (models_db['NA_der'] == 0)]
#                       & \
#              (models_db['outlier_bn_izq'] == 0) & (models_db['outlier_bn_der'] == 0) & \
#              (models_db['outlier_hs_izq'] == 0) & (models_db['outlier_hs_der'] == 0) & \
#              (models_db['outlier_hlf_izq'] <= 1) & (models_db['outlier_hlf_der'] <= 1)]

models_db.shape[0] + outliers.shape[0]

outliers.shape

# models_db.head()
models_db.shape

models_db.iloc[:,:40].columns

outliers.loc[:,['ID_izq','ID_der',
                'NA_izq','NA_der',
                'outlier_bn_izq','outlier_bn_der',
                'outlier_hs_izq','outlier_hs_der',
                'outlier_hlf_izq','outlier_hlf_der']].head()

users = models_db[['question', 'choice', 'user', 'age', 'gender', 'education', 'transport']]
# users = models_db[['question', 'choice']]
users.head()

data = pd.DataFrame()

# data['N_Bicyclist'] = models_db['Bicyclist_N_izq'] - models_db['Bicyclist_N_der']
# data['S_Bicyclist'] = models_db['Bicyclist_S_izq'] - models_db['Bicyclist_S_der']
# data['N_Building'] = models_db['Building_N_izq'] - models_db['Building_N_der']
# data['S_Building'] = models_db['Building_S_izq'] - models_db['Building_S_der']
# data['N_Car'] = models_db['Car_N_izq'] - models_db['Car_N_der']
# data['S_Car'] = models_db['Car_S_izq'] - models_db['Car_S_der']
# data['N_Fence'] = models_db['Fence_N_izq'] - models_db['Fence_N_der']
# data['S_Fence'] = models_db['Fence_S_izq'] - models_db['Fence_S_der']
# data['N_Pavement'] = models_db['Pavement_N_izq'] - models_db['Pavement_N_der']
# data['S_Pavement'] = models_db['Pavement_S_izq'] - models_db['Pavement_S_der']
# data['N_Pedestrian'] = models_db['Pedestrian_N_izq'] - models_db['Pedestrian_N_der']
# data['S_Pedestrian'] = models_db['Pedestrian_S_izq'] - models_db['Pedestrian_S_der']
# data['N_Pole'] = models_db['Pole_N_izq'] - models_db['Pole_N_der']
# data['S_Pole'] = models_db['Pole_S_izq'] - models_db['Pole_S_der']
# data['N_Road'] = models_db['Road_N_izq'] - models_db['Road_N_der']
# data['S_Road'] = models_db['Road_S_izq'] - models_db['Road_S_der']
# data['N_SignSymbol'] = models_db['SignSymbol_N_izq'] - models_db['SignSymbol_N_der']
# data['S_SignSymbol'] = models_db['SignSymbol_S_izq'] - models_db['SignSymbol_S_der']
# data['N_Sky'] = models_db['Sky_N_izq'] - models_db['Sky_N_der']
# data['S_Sky'] = models_db['Sky_S_izq'] - models_db['Sky_S_der']
# data['N_Tree'] = models_db['Tree_N_izq'] - models_db['Tree_N_der']
# data['S_Tree'] = models_db['Tree_S_izq'] - models_db['Tree_S_der']
data['N_Bicyclist'] = 0.1 * np.log(models_db['Bicyclist_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Bicyclist_N_izq'] * 100 + 1)
data['S_Bicyclist'] = 0.1 * np.log(models_db['Bicyclist_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Bicyclist_S_izq'] * 100 + 1)
data['N_Building'] = 0.1 * np.log(models_db['Building_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Building_N_izq'] * 100 + 1)
data['S_Building'] = 0.1 * np.log(models_db['Building_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Building_S_izq'] * 100 + 1)
data['N_Car'] = 0.1 * np.log(models_db['Car_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Car_N_izq'] * 100 + 1)
data['S_Car'] = 0.1 * np.log(models_db['Car_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Car_S_izq'] * 100 + 1)
data['N_Fence'] = 0.1 * np.log(models_db['Fence_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Fence_N_izq'] * 100 + 1)
data['S_Fence'] = 0.1 * np.log(models_db['Fence_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Fence_S_izq'] * 100 + 1)
data['N_Pavement'] = 0.1 * np.log(models_db['Pavement_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Pavement_N_izq'] * 100 + 1)
data['S_Pavement'] = 0.1 * np.log(models_db['Pavement_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Pavement_S_izq'] * 100 + 1)
data['N_Pedestrian'] = 0.1 * np.log(models_db['Pedestrian_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Pedestrian_N_izq'] * 100 + 1)
data['S_Pedestrian'] = 0.1 * np.log(models_db['Pedestrian_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Pedestrian_S_izq'] * 100 + 1)
data['N_Pole'] = 0.1 * np.log(models_db['Pole_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Pole_N_izq'] * 100 + 1)
data['S_Pole'] = 0.1 * np.log(models_db['Pole_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Pole_S_izq'] * 100 + 1)
data['N_Road'] = 0.1 * np.log(models_db['Road_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Road_N_izq'] * 100 + 1)
data['S_Road'] = 0.1 * np.log(models_db['Road_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Road_S_izq'] * 100 + 1)
data['N_SignSymbol'] = 0.1 * np.log(models_db['SignSymbol_N_der'] * 100 + 1) - 0.1 * np.log(models_db['SignSymbol_N_izq'] * 100 + 1)
data['S_SignSymbol'] = 0.1 * np.log(models_db['SignSymbol_S_der'] * 100 + 1) - 0.1 * np.log(models_db['SignSymbol_S_izq'] * 100 + 1)
data['N_Sky'] = 0.1 * np.log(models_db['Sky_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Sky_N_izq'] * 100 + 1)
data['S_Sky'] = 0.1 * np.log(models_db['Sky_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Sky_S_izq'] * 100 + 1)
data['N_Tree'] = 0.1 * np.log(models_db['Tree_N_der'] * 100 + 1) - 0.1 * np.log(models_db['Tree_N_izq'] * 100 + 1)
data['S_Tree'] = 0.1 * np.log(models_db['Tree_S_der'] * 100 + 1) - 0.1 * np.log(models_db['Tree_S_izq'] * 100 + 1)

# data = (data-data.mean())/data.std()
# data = data*10

data['blobs'] = models_db['blobs_der'] - models_db['blobs_izq']
data['edges'] = models_db['edges_der'] - models_db['edges_izq']
data['mean_bn'] = models_db['mean_bn_der'] - models_db['mean_bn_izq']
data['mean_h'] = models_db['mean_h_der'] - models_db['mean_h_izq']
data['mean_l'] = models_db['mean_l_der'] - models_db['mean_l_izq']
data['mean_s'] = models_db['mean_s_der'] - models_db['mean_s_izq']
data['stdev_bn'] = models_db['stdev_bn_der'] - models_db['stdev_bn_izq']
data['stdev_h'] = models_db['stdev_h_der'] - models_db['stdev_h_izq']
data['stdev_l'] = models_db['stdev_l_der'] - models_db['stdev_l_izq']
data['stdev_s'] = models_db['stdev_s_der'] - models_db['stdev_s_izq']

data['OD_bench'] = models_db['od_bench_der'] - models_db['od_bench_izq']
data['OD_bicycle_der'] = models_db['od_bicycle_der']
data['OD_bicycle_izq'] = models_db['od_bicycle_izq']
data['OD_boat'] = models_db['od_boat_der'] - models_db['od_boat_izq']
data['OD_bus_der'] = models_db['od_bus_der'] 
data['OD_bus_izq'] = models_db['od_bus_izq']
data['OD_car_der'] = models_db['od_car_der'] 
data['OD_car_izq'] = models_db['od_car_izq']
data['OD_chair'] = models_db['od_chair_der'] - models_db['od_chair_izq']
data['OD_fire_hydrant'] = models_db['od_fire_hydrant_der'] - models_db['od_fire_hydrant_izq']
data['OD_motorcycle_der'] = models_db['od_motorcycle_der']
data['OD_motorcycle_izq'] = models_db['od_motorcycle_izq']
data['OD_person_der'] = models_db['od_person_der']
data['OD_person_izq'] = models_db['od_person_izq']
data['OD_potted_plant'] = models_db['od_potted_plant_der'] - models_db['od_potted_plant_izq']
data['OD_stop_sign'] = models_db['od_stop_sign_der'] - models_db['od_stop_sign_izq']
data['OD_traffic_light_der'] = models_db['od_traffic_light_der'] 
data['OD_traffic_light_izq'] = models_db['od_traffic_light_izq']
data['OD_train'] = models_db['od_train_der'] - models_db['od_train_izq']
data['OD_truck_der'] = models_db['od_truck_der'] 
data['OD_truck_izq'] = models_db['od_truck_izq']
data['OD_umbrella'] = models_db['od_umbrella_der'] - models_db['od_umbrella_izq']

data = data.fillna(0)

data.shape
data.describe().T

data.head()

all_data = users.join(data)

all_data.loc[(all_data['question']==1),'OD_car_der'].describe().T

# all_data.shape
all_data.head()

all_data.to_csv('./../00 Biogeme Input/log_data-sin_filtrar.dat', sep='\t', header=True, index=False, doublequote=True, decimal='.')
