==> analysis/02 Preparar data para aplicacion.ipynb <==
import pandas as pd
import numpy as np
import os

# os.chdir('./../00 Insumos/01 PP')
os.chdir('./../00 Insumos/02 Stgo')
os.listdir()

HLF = pd.read_csv('hlfeat.csv').sort_values('ID').reset_index(drop=True) #, index_col='ID')
LLF = pd.read_csv('llfeat.csv').sort_values('ID').reset_index(drop=True) #,index_col='ID')
# nans = pd.read_csv('nans.csv',index_col='ID')
od = pd.read_csv('OD.csv').sort_values('ID').reset_index(drop=True) #,index_col='ID')

# HLF = HLF.merge(od,how='left',on='ID')
# features = pd.concat([HLF,LLF],axis=1)

features = pd.concat([LLF, HLF, od], axis=1)

features.to_csv('all_features_imagenes.csv', index=False)

# features = pd.read_csv('all_features_imagenes.csv')
features



features.columns

data = pd.DataFrame()

data['N_Bicyclist'] = 0.1 * np.log(features['Bicyclist_N'] * 100 + 1)
data['S_Bicyclist'] = 0.1 * np.log(features['Bicyclist_S'] * 100 + 1)
data['N_Building'] = 0.1 * np.log(features['Building_N'] * 100 + 1)
data['S_Building'] = 0.1 * np.log(features['Building_S'] * 100 + 1) 
data['N_Car'] = 0.1 * np.log(features['Car_N'] * 100 + 1)   
data['S_Car'] = 0.1 * np.log(features['Car_S'] * 100 + 1)            
data['N_Fence'] = 0.1 * np.log(features['Fence_N'] * 100 + 1)         
data['S_Fence'] = 0.1 * np.log(features['Fence_S'] * 100 + 1)         
data['N_Pavement'] = 0.1 * np.log(features['Pavement_N'] * 100 + 1)   
data['S_Pavement'] = 0.1 * np.log(features['Pavement_S'] * 100 + 1)   
data['N_Pedestrian'] = 0.1 * np.log(features['Pedestrian_N'] * 100 + 1)
data['S_Pedestrian'] = 0.1 * np.log(features['Pedestrian_S'] * 100 + 1) 
data['N_Pole'] = 0.1 * np.log(features['Pole_N'] * 100 + 1)
data['S_Pole'] = 0.1 * np.log(features['Pole_S'] * 100 + 1)
data['N_Road'] = 0.1 * np.log(features['Road_N'] * 100 + 1)
data['S_Road'] = 0.1 * np.log(features['Road_S'] * 100 + 1)
data['N_SignSymbol'] = 0.1 * np.log(features['SignSymbol_N'] * 100 + 1)
data['S_SignSymbol'] = 0.1 * np.log(features['SignSymbol_S'] * 100 + 1)
data['N_Sky'] = 0.1 * np.log(features['Sky_N'] * 100 + 1)
data['S_Sky'] = 0.1 * np.log(features['Sky_S'] * 100 + 1)
data['N_Tree'] = 0.1 * np.log(features['Tree_N'] * 100 + 1)
data['S_Tree'] = 0.1 * np.log(features['Tree_S'] * 100 + 1)

data['blobs'] = features['blobs']
data['edges'] = features['edges']
data['mean_bn'] = features['mean_bn']
data['mean_h'] = features['mean_h']
data['mean_l'] = features['mean_l']
data['mean_s'] = features['mean_s']
data['stdev_bn'] = features['stdev_bn']
data['stdev_h'] = features['stdev_h']
data['stdev_l'] = features['stdev_l']
data['stdev_s'] = features['stdev_s']

data['OD_bench'] = features['od_bench']
data['OD_bicycle_der'] = features['od_bicycle']
data['OD_bicycle_izq'] = 0
data['OD_boat'] = features['od_boat']
data['OD_bus_der'] = features['od_bus'] 
data['OD_bus_izq'] = 0
data['OD_car_der'] = features['od_car'] 
data['OD_car_izq'] = 0
data['OD_chair'] = features['od_chair']
data['OD_fire_hydrant'] = features['od_fire_hydrant']
data['OD_motorcycle_der'] = features['od_motorcycle']
data['OD_motorcycle_izq'] = 0
data['OD_person_der'] = features['od_person']
data['OD_person_izq'] = 0
data['OD_potted_plant'] = features['od_potted_plant']
data['OD_stop_sign'] = features['od_stop_sign']
data['OD_traffic_light_der'] = features['od_traffic_light'] 
data['OD_traffic_light_izq'] = 0
data['OD_train'] = features['od_train']
data['OD_truck_der'] = features['od_truck'] 
data['OD_truck_izq'] = 0
data['OD_umbrella'] = features['od_umbrella']

data = data.fillna(0)

data.to_csv('data_aplicacion.csv')
