==> analysis/06 Aplicacion Live.ipynb <==
import pandas as pd
import os

import biogeme.database as db
import biogeme.biogeme as bio
import biogeme.distributions as dist
import biogeme.models as models
import biogeme.results as res

data = pd.read_csv('./../00 Insumos/02 Stgo/data_aplicacion.csv')

data['transport'] = 0
data['gender'] = 0
data['choice'] = 1
database = db.Database("database",data)
from headers import *

B_mean_h = Beta('LLF: mean_h', 0, -100, 100, 0)
B_mean_l = Beta('LLF: mean_l', 0, -100, 100, 0)
B_mean_s = Beta('LLF: mean_s', 0, -100, 100, 0)
B_mean_bn = Beta('LLF: mean_bn', 0, -100, 100, 1)  # NS
B_stdev_h = Beta('LLF: stdev_h', 0, -100, 100, 0)
B_stdev_l = Beta('LLF: stdev_l', 0, -100, 100, 0)
B_stdev_s = Beta('LLF: stdev_s', 0, -100, 100, 0)
B_stdev_bn = Beta('LLF: stdev_bn', 0, -100, 100, 1)  # NS

B_edges = Beta('LLF: edges', 0, -100, 100, 0)
B_blobs = Beta('LLF: blobs', 0, -100, 100, 1)


B_Bicyclist_N = Beta('HLF N: Bicyclist', 0, -100, 100, 0) 
B_Building_N = Beta('HLF N: Building', 0, -100, 100, 1)
B_Car_N = Beta('HLF N: Car', 0, -100, 100, 0) 
B_Fence_N = Beta('HLF N: Fence', 0, -100, 100, 0) 
B_Pavement_N = Beta('HLF N: Pavement', 0, -100, 100, 1)  # NS 0
B_Pedestrian_N = Beta('HLF N: Pedestrian', 0, -100, 100, 1)
B_Pole_N = Beta('HLF N: Pole', 0, -100, 100, 0) 
B_Road_N = Beta('HLF N: Road', 0, -100, 100, 0)  # NS 0
B_SignSymbol_N = Beta('HLF N: SignSymbol', 0, -100, 100, 0) 
B_Sky_N = Beta('HLF N: Sky', 0, -100, 100, 1)
B_Tree_N = Beta('HLF N: Tree', 0, -100, 100, 0)

B_Bicyclist_S = Beta('HLF S: Bicyclist', 0, -100, 100, 1) 
B_Building_S = Beta('HLF S: Building', 0, -100, 100, 1) 
B_Car_S = Beta('HLF S: Car', 0, -100, 100, 1)
B_Fence_S = Beta('HLF S: Fence', 0, -100, 100, 0) 
B_Pavement_S = Beta('HLF S: Pavement', 0, -100, 100, 1)
B_Pedestrian_S = Beta('HLF S: Pedestrian', 0, -100, 100, 1)
B_Pole_S = Beta('HLF S: Pole', 0, -100, 100, 0)
B_Road_S = Beta('HLF S: Road', 0, -100, 100, 0)
B_SignSymbol_S = Beta('HLF S: SignSymbol', 0, -100, 100, 1)
B_Sky_S = Beta('HLF S: Sky', 0, -100, 100, 1) # NS 0 
B_Tree_S = Beta('HLF S: Tree', 0, -100, 100, 0) 


D_bench = Beta('OD: bench', 0, -100, 100, 1)
D_boat = Beta('OD: boat', 0, -100, 100, 1)
D_umbrella = Beta('OD: umbrella', 0, -100, 100, 1)
D_chair = Beta('OD: chair', 0, -100, 100, 1) ## SI VA

D_train = Beta('OD: train', 0, -100, 100, 1)## SI VA
D_potted_plant = Beta('OD: potted_plant', 0, -100, 100, 1)
D_stop_sign = Beta('OD: stop_sign', 0, -100, 100, 1)
D_traffic_light = Beta('OD: traffic_light', 0, -100, 100, 1)

D_bus = Beta('OD: bus', 0, -100, 100, 1)
D_motorcycle = Beta('OD: motorcycle', 0, -100, 100, 1)

# D_person = Beta('OD: person', 0, -100, 100, 1)
# D_car = Beta('OD: car', 0, -100, 100, 1)
D_fire_hydrant = Beta('OD: fire_hydrant', 0, -100, 100, 1)

D_truck = Beta('OD: truck', 0, -100, 100, 0)
LAMBDA_truck = Beta('LAMBDA: truck',1,-5,5, 0)## SI VA

D_bike = Beta('OD: bike', 0, -100, 100, 1)
LAMBDA_bike = Beta('LAMBDA: bike',1,-5,5, 1)## SI VA

D_car = Beta('OD: car', 0, -100, 100, 0)## SI VA
LAMBDA_car = Beta('LAMBDA: car',1,-5,15, 0)## SI VA

D_person = Beta('OD: person', 0, -100, 100, 1)## SI VA
LAMBDA_person = Beta('LAMBDA: person',1,-15,5,1)## SI VA

D_light = Beta('OD: light', 0, -100, 100, 1)
LAMBDA_light = Beta('LAMBDA: light',1.0,0.0001,5,1)

# D_car_01 = Beta('OD: car 1', 0, -100, 100, 1)
# D_car_02 = Beta('OD: car 2', 0, -100, 100, 1)
# D_car_03 = Beta('OD: car 3', 0, -100, 100, 1)
# D_car_04 = Beta('OD: car 4', 0, -100, 100, 1)
# D_car_05 = Beta('OD: car 5', 0, -100, 100, 1)
# D_car_06 = Beta('OD: car 6', 0, -100, 100, 1)
# D_car_07 = Beta('OD: car 7', 0, -100, 100, 1)
# D_car_08 = Beta('OD: car 8', 0, -100, 100, 1)
# D_car_09 = Beta('OD: car 9', 0, -100, 100, 1)
# D_car_10 = Beta('OD: car 10', 0, -100, 100, 1)
# D_car_15 = Beta('OD: car 15', 0, -100, 100, 1)
# D_car_20 = Beta('OD: car 20', 0, -100, 100, 1)
# D_car_40 = Beta('OD: car 40+', 0, -100, 100, 1)


tau1	 = Beta('tau1',-1,None,0,0)
delta2	 = Beta('delta2',2,0,None,0)

tau2 = tau1 + delta2

G_Bicyclist_N = Beta('Female w/ HLF N: Bicyclist', 0, -100, 100, 1) 
G_Building_N = Beta('Female w/ HLF N: Building', 0, -100, 100, 1)
G_Car_N = Beta('Female w/ HLF N: Car', 0, -100, 100, 1) 
G_Fence_N = Beta('Female w/ HLF N: Fence', 0, -100, 100, 1) 
G_Pavement_N = Beta('Female w/ HLF N: Pavement', 0, -100, 100, 1)  # NS 0
G_Pedestrian_N = Beta('Female w/ HLF N: Pedestrian', 0, -100, 100, 0)
G_Pole_N = Beta('Female w/ HLF N: Pole', 0, -100, 100, 1) 
G_Road_N = Beta('Female w/ HLF N: Road', 0, -100, 100, 1)  # NS 0
G_SignSymbol_N = Beta('Female w/ HLF N: SignSymbol', 0, -100, 100, 0) 
G_Sky_N = Beta('Female w/ HLF N: Sky', 0, -100, 100, 1)
G_Tree_N = Beta('Female w/ HLF N: Tree', 0, -100, 100, 0)

G_Bicyclist_S = Beta('Female w/ HLF S: Bicyclist', 0, -100, 100, 1) 
G_Building_S = Beta('Female w/ HLF S: Building', 0, -100, 100, 0) 
G_Car_S = Beta('Female w/ HLF S: Car', 0, -100, 100, 1)
G_Fence_S = Beta('Female w/ HLF S: Fence', 0, -100, 100, 1) 
G_Pavement_S = Beta('Female w/ HLF S: Pavement', 0, -100, 100, 1)
G_Pedestrian_S = Beta('Female w/ HLF S: Pedestrian', 0, -100, 100, 1)
G_Pole_S = Beta('Female w/ HLF S: Pole', 0, -100, 100, 1)
G_Road_S = Beta('Female w/ HLF S: Road', 0, -100, 100, 1)
G_SignSymbol_S = Beta('Female w/ HLF S: SignSymbol', 0, -100, 100, 1)
G_Sky_S = Beta('Female w/ HLF S: Sky', 0, -100, 100, 1) # NS 0 
G_Tree_S = Beta('Female w/ HLF S: Tree', 0, -100, 100, 1) 

G_bench = Beta('Female w/ OD: bench', 0, -100, 100, 1)
G_boat = Beta('Female w/ OD: boat', 0, -100, 100, 0)
G_umbrella = Beta('Female w/ OD: umbrella', 0, -100, 100, 1)
G_chair = Beta('Female w/ OD: chair', 0, -100, 100, 0) ## SI VA

G_train = Beta('Female w/ OD: train', 0, -100, 100, 1)## SI VA
G_potted_plant = Beta('Female w/ OD: potted_plant', 0, -100, 100, 1)
G_stop_sign = Beta('Female w/ OD: stop_sign', 0, -100, 100, 1)
G_bus = Beta('Female w/ OD: bus', 0, -100, 100, 1)
G_motorcycle = Beta('Female w/ OD: motorcycle', 0, -100, 100, 1)

# G_person = Beta('Female w/ OD: person', 0, -100, 100, 0)
# G_car = Beta('Female w/ OD: car', 0, -100, 100, 0)
G_fire_hydrant = Beta('Female w/ OD: fire_hydrant', 0, -100, 100, 1)

G_truck = Beta('Female w/ OD: truck', 0, -100, 100, 1)
GAMMA_truck = Beta('Female w/ GAMMA: truck',1,-5,5, 1)## SI VA

G_bike = Beta('Female w/ OD: bike', 0, -100, 100, 1)
GAMMA_bike = Beta('Female w/ GAMMA: bike',1.0,-5,5, 1)## SI VA

G_car = Beta('Female w/ OD: car', 0, -100, 100, 1)## SI VA
GAMMA_car = Beta('Female w/ GAMMA: car',1.0,0.1,15, 1)## SI VA

G_person = Beta('Female w/ OD: person', 0, -100, 100, 1)## SI VA
GAMMA_person = Beta('Female w/ GAMMA: person',1.0,0.0001,5,1)## SI VA

G_light = Beta('Female w/ OD: light', 0, -100, 100, 1)
GAMMA_light = Beta('Female w/ GAMMA: light',1.0,0.0001,5,1)

PED_Building_N = Beta('PED w/ HLF N: Building', 0, -100, 100, 1)
PED_Bicyclist_S = Beta('PED w/ HLF S: Bicyclist', 0, -100, 100, 1) 
PED_car = Beta('PED w/ OD: car', 0, -100, 100, 1)## SI VA
PED_stop_sign = Beta('PED w/ OD: stop_sign', 0, -100, 100, 1)

BIKE_Bicyclist_N = Beta('BIKE w/ HLF N: Bicyclist', 0, -100, 100, 1) 
BIKE_Building_N = Beta('BIKE w/ HLF N: Building', 0, -100, 100, 1)
BIKE_Car_N = Beta('BIKE w/ HLF N: Car', 0, -100, 100, 0) 
BIKE_Pole_N = Beta('BIKE w/ HLF N: Pole', 0, -100, 100, 0) 
BIKE_SignSymbol_N = Beta('BIKE w/ HLF N: SignSymbol', 0, -100, 100, 0)
BIKE_Car_S = Beta('BIKE w/ HLF S: Car', 0, -100, 100, 0)
BIKE_Pole_S = Beta('BIKE w/ HLF S: Pole', 0, -100, 100, 1)
BIKE_Road_S = Beta('BIKE w/ HLF S: Road', 0, -100, 100, 0)
BIKE_SignSymbol_S = Beta('BIKE w/ HLF S: SignSymbol', 0, -100, 100, 0)
BIKE_bench = Beta('BIKE w/ OD: bench', 0, -100, 100, 0)
BIKE_car = Beta('BIKE w/ OD: car', 0, -100, 100, 0)## SI VA
BIKE_fire_hydrant = Beta('BIKE w/ OD: fire_hydrant', 0, -100, 100, 1)
BIKE_light = Beta('BIKE w/ OD: light', 0, -100, 100, 1)

PT_Fence_N = Beta('PT w/ HLF N: Fence', 0, -100, 100, 1)
PT_Pole_S = Beta('PT w/ HLF S: Pole', 0, -100, 100, 1)
PT_bench = Beta('PT w/ OD: bench', 0, -100, 100, 1)
PT_bike = Beta('PT w/ OD: bike', 0, -100, 100, 1)

CAR_Fence_N = Beta('CAR w/ HLF N: Fence', 0, -100, 100, 0) 
CAR_Tree_N = Beta('CAR w/ HLF N: Tree', 0, -100, 100, 1)
CAR_car = Beta('CAR w/ OD: car', 0, -100, 100, 0)## SI VA

M12_Bicyclist_N = Beta('PED n BIKE w/ HLF N: Bicyclist', 0, -100, 100, 0) 
M12_Building_N = Beta('PED n BIKE w/ HLF N: Building', 0, -100, 100, 0)
M12_Pedestrian_N = Beta('PED n BIKE w/ HLF N: Pedestrian', 0, -100, 100, 0)
M12_SignSymbol_N = Beta('PED n BIKE w/ HLF N: SignSymbol', 0, -100, 100, 1) 
M12_Bicyclist_S = Beta('PED n BIKE w/ HLF S: Bicyclist', 0, -100, 100, 1) 
M12_Pole_S = Beta('PED n BIKE w/ HLF S: Pole', 0, -100, 100, 0)
M12_car = Beta('PED n BIKE w/ OD: car', 0, -100, 100, 1)## SI VA
M12_fire_hydrant = Beta('PED n BIKE w/ OD: fire_hydrant', 0, -100, 100, 0)
M12_light = Beta('PED n BIKE w/ OD: light', 0, -100, 100, 0)



U = B_mean_h * mean_h + \
     B_mean_l * mean_l + \
     B_mean_s * mean_s + \
     B_mean_bn * mean_bn + \
     B_stdev_h * stdev_h + \
     B_stdev_l * stdev_l + \
     B_stdev_s * stdev_s + \
     B_stdev_bn * stdev_bn + \
     B_edges * edges + \
     B_blobs * blobs + \
     B_Bicyclist_N * N_Bicyclist + \
     B_Building_N * N_Building + \
     B_Car_N * N_Car + \
     B_Fence_N * N_Fence + \
     B_Pavement_N * N_Pavement + \
     B_Pedestrian_N * N_Pedestrian + \
     B_Pole_N * N_Pole + \
     B_Road_N * N_Road + \
     B_SignSymbol_N * N_SignSymbol + \
     B_Sky_N * N_Sky + \
     B_Tree_N * N_Tree + \
     B_Bicyclist_S * S_Bicyclist + \
     B_Building_S * S_Building + \
     B_Car_S * S_Car + \
     B_Fence_S * S_Fence + \
     B_Pavement_S * S_Pavement + \
     B_Pedestrian_S * S_Pedestrian + \
     B_Pole_S * S_Pole + \
     B_Road_S * S_Road + \
     B_SignSymbol_S * S_SignSymbol + \
     B_Sky_S * S_Sky + \
     B_Tree_S * S_Tree + \
     D_bench * OD_bench + \
     D_chair * OD_chair + \
     D_fire_hydrant * OD_fire_hydrant + \
     D_potted_plant * OD_potted_plant + \
     D_stop_sign * OD_stop_sign + \
     D_train * OD_train + \
     D_umbrella * OD_umbrella + \
     D_bus * ( OD_bus_der - OD_bus_izq )+ \
     D_truck * (OD_truck_der) ** LAMBDA_truck - \
     D_truck * (OD_truck_izq) ** LAMBDA_truck + \
     D_person * (OD_person_der) ** LAMBDA_person - \
     D_person * (OD_person_izq) ** LAMBDA_person + \
     D_car * (OD_car_der)**LAMBDA_car - \
     D_car * (OD_car_izq)**LAMBDA_car + \
     D_light * (OD_traffic_light_der) ** LAMBDA_light - \
     D_light * (OD_traffic_light_izq) ** LAMBDA_light + \
     D_bike * (OD_bicycle_der) ** LAMBDA_bike - \
     D_bike * (OD_bicycle_izq) ** LAMBDA_bike + \
     gender * G_Bicyclist_N * N_Bicyclist + \
     gender * G_Building_N * N_Building + \
     gender * G_Car_N * N_Car + \
     gender * G_Fence_N * N_Fence + \
     gender * G_Pavement_N * N_Pavement + \
     gender * G_Pedestrian_N * N_Pedestrian + \
     gender * G_Pole_N * N_Pole + \
     gender * G_Road_N * N_Road + \
     gender * G_SignSymbol_N * N_SignSymbol + \
     gender * G_Sky_N * N_Sky + \
     gender * G_Tree_N * N_Tree + \
     gender * G_Bicyclist_S * S_Bicyclist + \
     gender * G_Building_S * S_Building + \
     gender * G_Car_S * S_Car + \
     gender * G_Fence_S * S_Fence + \
     gender * G_Pavement_S * S_Pavement + \
     gender * G_Pedestrian_S * S_Pedestrian + \
     gender * G_Pole_S * S_Pole + \
     gender * G_Road_S * S_Road + \
     gender * G_SignSymbol_S * S_SignSymbol + \
     gender * G_Sky_S * S_Sky + \
     gender * G_Tree_S * S_Tree + \
     gender * G_bench * OD_bench + \
     gender * G_chair * OD_chair + \
     gender * G_fire_hydrant * OD_fire_hydrant + \
     gender * G_potted_plant * OD_potted_plant + \
     gender * G_stop_sign * OD_stop_sign + \
     gender * G_train * OD_train + \
     gender * G_umbrella * OD_umbrella + \
     gender * G_bus * ( OD_bus_der - OD_bus_izq )+ \
     gender * G_truck * (OD_truck_der) ** LAMBDA_truck - \
     gender * G_truck * (OD_truck_izq) ** LAMBDA_truck + \
     gender * G_person * (OD_person_der)** LAMBDA_person - \
     gender * G_person * (OD_person_izq)** LAMBDA_person + \
     gender * G_car * (OD_car_der)** LAMBDA_car - \
     gender * G_car * (OD_car_izq)** LAMBDA_car + \
     gender * G_light * (OD_traffic_light_der) ** LAMBDA_light - \
     gender * G_light * (OD_traffic_light_izq) ** LAMBDA_light + \
     gender * G_bike * (OD_bicycle_der) ** LAMBDA_bike - \
     gender * G_bike * (OD_bicycle_izq) ** LAMBDA_bike + \
     (transport == 1) * PED_Building_N * N_Building + \
     (transport == 1) * PED_Bicyclist_S * S_Bicyclist + \
     (transport == 1) * PED_car * (OD_car_der)**LAMBDA_car - \
     (transport == 1) * PED_car * (OD_car_izq)**LAMBDA_car + \
     (transport == 1) * PED_stop_sign * OD_stop_sign + \
     (transport == 2) * BIKE_Bicyclist_N * N_Bicyclist + \
     (transport == 2) * BIKE_Building_N * N_Building + \
     (transport == 2) * BIKE_Car_N * N_Car + \
     (transport == 2) * BIKE_Pole_N * N_Pole + \
     (transport == 2) * BIKE_SignSymbol_N * N_SignSymbol + \
     (transport == 2) * BIKE_Car_S * S_Car + \
     (transport == 2) * BIKE_Pole_S * S_Pole + \
     (transport == 2) * BIKE_Road_S * S_Road + \
     (transport == 2) * BIKE_SignSymbol_S * S_SignSymbol + \
     (transport == 2) * BIKE_bench * OD_bench + \
     (transport == 2) * BIKE_fire_hydrant * OD_fire_hydrant + \
     (transport == 2) * BIKE_car * (OD_car_der)**LAMBDA_car - \
     (transport == 2) * BIKE_car * (OD_car_izq)**LAMBDA_car + \
     (transport == 2) * BIKE_light * (OD_traffic_light_der) ** LAMBDA_light - \
     (transport == 2) * BIKE_light * (OD_traffic_light_izq) ** LAMBDA_light + \
     (transport == 3) * PT_Fence_N * N_Fence + \
     (transport == 3) * PT_Pole_S * S_Pole + \
     (transport == 3) * PT_bench * OD_bench + \
     (transport == 3) * PT_bike * (OD_bicycle_der) ** LAMBDA_bike - \
     (transport == 3) * PT_bike * (OD_bicycle_izq) ** LAMBDA_bike + \
     (transport == 4) * CAR_Fence_N * N_Fence + \
     (transport == 4) * CAR_Tree_N * N_Tree + \
     (transport == 4) * CAR_car * (OD_car_der)**LAMBDA_car - \
     (transport == 4) * CAR_car * (OD_car_izq)**LAMBDA_car + \
     ((transport == 1) + (transport == 2)) * M12_Bicyclist_N * N_Bicyclist + \
     ((transport == 1) + (transport == 2)) * M12_Building_N * N_Building + \
     ((transport == 1) + (transport == 2)) * M12_Pedestrian_N * N_Pedestrian + \
     ((transport == 1) + (transport == 2)) * M12_SignSymbol_N * N_SignSymbol + \
     ((transport == 1) + (transport == 2)) * M12_Bicyclist_S * S_Bicyclist + \
     ((transport == 1) + (transport == 2)) * M12_Pole_S * S_Pole + \
     ((transport == 1) + (transport == 2)) * M12_light * (OD_traffic_light_der) ** LAMBDA_light - \
     ((transport == 1) + (transport == 2)) * M12_light * (OD_traffic_light_izq) ** LAMBDA_light + \
     ((transport == 1) + (transport == 2)) * M12_fire_hydrant * OD_fire_hydrant + \
     ((transport == 1) + (transport == 2)) * M12_car * (OD_car_der)**LAMBDA_car - \
     ((transport == 1) + (transport == 2)) * M12_car * (OD_car_izq)**LAMBDA_car


#      ((transport == 1) + (transport == 2)) * VS_Bicyclist_N * N_Bicyclist + \
#      ((transport == 1) + (transport == 2)) * VS_Building_N * N_Building + \
#      ((transport == 1) + (transport == 2)) * VS_Car_N * N_Car + \
#      ((transport == 1) + (transport == 2)) * VS_Fence_N * N_Fence + \
#      ((transport == 1) + (transport == 2)) * VS_Pavement_N * N_Pavement + \
#      ((transport == 1) + (transport == 2)) * VS_Pedestrian_N * N_Pedestrian + \
#      ((transport == 1) + (transport == 2)) * VS_Pole_N * N_Pole + \
#      ((transport == 1) + (transport == 2)) * VS_Road_N * N_Road + \
#      ((transport == 1) + (transport == 2)) * VS_SignSymbol_N * N_SignSymbol + \
#      ((transport == 1) + (transport == 2)) * VS_Sky_N * N_Sky + \
#      ((transport == 1) + (transport == 2)) * VS_Tree_N * N_Tree + \
#      ((transport == 1) + (transport == 2)) * VS_Bicyclist_S * S_Bicyclist + \
#      ((transport == 1) + (transport == 2)) * VS_Building_S * S_Building + \
#      ((transport == 1) + (transport == 2)) * VS_Car_S * S_Car + \
#      ((transport == 1) + (transport == 2)) * VS_Fence_S * S_Fence + \
#      ((transport == 1) + (transport == 2)) * VS_Pavement_S * S_Pavement + \
#      ((transport == 1) + (transport == 2)) * VS_Pedestrian_S * S_Pedestrian + \
#      ((transport == 1) + (transport == 2)) * VS_Pole_S * S_Pole + \
#      ((transport == 1) + (transport == 2)) * VS_Road_S * S_Road + \
#      ((transport == 1) + (transport == 2)) * VS_SignSymbol_S * S_SignSymbol + \
#      ((transport == 1) + (transport == 2)) * VS_Sky_S * S_Sky + \
#      ((transport == 1) + (transport == 2)) * VS_Tree_S * S_Tree + \
#      ((transport == 1) + (transport == 2)) * VS_bench * OD_bench + \
#      ((transport == 1) + (transport == 2)) * VS_chair * OD_chair + \
#      ((transport == 1) + (transport == 2)) * VS_fire_hydrant * OD_fire_hydrant + \
#      ((transport == 1) + (transport == 2)) * VS_potted_plant * OD_potted_plant + \
#      ((transport == 1) + (transport == 2)) * VS_stop_sign * OD_stop_sign + \
#      ((transport == 1) + (transport == 2)) * VS_train * OD_train + \
#      ((transport == 1) + (transport == 2)) * VS_umbrella * OD_umbrella + \
#      ((transport == 1) + (transport == 2)) * VS_bus * ( OD_bus_der - OD_bus_izq )+ \
#      ((transport == 1) + (transport == 2)) * VS_truck * (OD_truck_der) ** LAMBDA_truck - \
#      ((transport == 1) + (transport == 2)) * VS_truck * (OD_truck_izq) ** LAMBDA_truck + \
#      ((transport == 1) + (transport == 2)) * VS_person * (OD_person_der)**LAMBDA_person - \
#      ((transport == 1) + (transport == 2)) * VS_person * (OD_person_izq)**LAMBDA_person + \
#      ((transport == 1) + (transport == 2)) * VS_car * (OD_car_der)**LAMBDA_car - \
#      ((transport == 1) + (transport == 2)) * VS_car * (OD_car_izq)**LAMBDA_car + \
#      ((transport == 1) + (transport == 2)) * VS_light * (OD_traffic_light_der) ** LAMBDA_light - \
#      ((transport == 1) + (transport == 2)) * VS_light * (OD_traffic_light_izq) ** LAMBDA_light + \
#      ((transport == 1) + (transport == 2)) * VS_bike * (OD_bicycle_der) ** LAMBDA_bike - \
#      ((transport == 1) + (transport == 2)) * VS_bike * (OD_bicycle_izq) ** LAMBDA_bike 

# gender * G_Pavement_S * ( S_Pavement + S_Road ) + \




P1 = 1-dist.logisticcdf(U-tau1)
P2 = dist.logisticcdf(U-tau2)
Pe = dist.logisticcdf(U-tau1)- dist.logisticcdf(U-tau2)

ChoiceProba = {
    1: P1,
    3: Pe,
    2: P2}

# bioNormalCdf
# logisticcdf
choice = 1

logprob = log(Elem(ChoiceProba,choice))
formulas = {'loglike':logprob,
           'Prob 1':P1,
           'Prob =':Pe,
           'Prob 2':P2,
           'Utilidad':U}

model = './../04 Modelos/__ Definitivos/12 live - VS.pickle'
results = res.bioResults(pickleFile = model)

betas = dict()
for i in range(len(results.data.betaValues)):
    betas[results.data.betaNames[i]] = results.data.betaValues[i]

df_eval = pd.read_csv('./../00 Insumos/02 Stgo/data_aplicacion.csv')
# features = pd.read_csv('./../../03 Respuestas/new_features_imagenes.csv')

betas

df_eval['transport'] = 0
df_eval['gender'] = 0
df_eval['choice'] = 1

database = db.Database("database",df_eval)

biogeme = bio.BIOGEME(database,formulas,numberOfThreads=8)
sim_male = biogeme.simulate(betas)

df_eval.loc[:,'gender'] = 1

database = db.Database("database",df_eval)

biogeme = bio.BIOGEME(database,formulas,numberOfThreads=8)
sim_female = biogeme.simulate(betas)

df_eval.loc[:,'transport'] = 1
df_eval.loc[:,'gender'] = 0

database = db.Database("database",df_eval)

biogeme = bio.BIOGEME(database,formulas,numberOfThreads=8)
sim_pedestrian = biogeme.simulate(betas)

df_eval.loc[:,'transport'] = 2

database = db.Database("database",df_eval)

biogeme = bio.BIOGEME(database,formulas,numberOfThreads=8)
sim_bike = biogeme.simulate(betas)

df_eval.loc[:,'transport'] = 3
database = db.Database("database",df_eval)

biogeme = bio.BIOGEME(database,formulas,numberOfThreads=8)
sim_PT = biogeme.simulate(betas)

df_eval.loc[:,'transport'] = 4

database = db.Database("database",df_eval)

biogeme = bio.BIOGEME(database,formulas,numberOfThreads=8)
sim_car = biogeme.simulate(betas)

score = pd.DataFrame(sim_male['Utilidad'])
score = score.join(sim_female['Utilidad'],how='left',rsuffix='_female')
score = score.join(sim_pedestrian['Utilidad'],how='left',rsuffix='_pedestrian')
score = score.join(sim_bike['Utilidad'],how='left',rsuffix='_bike')
score = score.join(sim_PT['Utilidad'],how='left',rsuffix='_PT')
score = score.join(sim_car['Utilidad'],how='left',rsuffix='_car')

score['D_female'] = score['Utilidad'] - score['Utilidad_female']
score['D_pedestrian'] = score['Utilidad'] - score['Utilidad_pedestrian']
score['D_bike'] = score['Utilidad'] - score['Utilidad_bike']
score['D_PT'] = score['Utilidad'] - score['Utilidad_PT']
score['D_car'] = score['Utilidad'] - score['Utilidad_car']

features = pd.read_csv('./../00 Insumos/02 Stgo/all_features_imagenes.csv')

perception = pd.concat([features['ID'],score.iloc[:,0],score.iloc[:,-5:]], axis=1)

perception.corr()

perception.to_csv('puntajes_live.csv')

# import matplotlib.pyplot as plt
# import requests
# from PIL import Image
# from io import BytesIO
# from io import StringIO

url_base = 'http://wekun.ing.puc.cl/static/img/'

for image_id in diff.sort_values('DIF').iloc[:5,[0]].values:
    url = url_base + image_id[0] + '.jpg'
    response = requests.get(url)
    image = Image.open(BytesIO(response.content))
    plt.imshow(image)
    plt.show()
