==> analysis/99 track ip.ipynb <==

import os
import pandas as pd
# from datetime import datetime
# import sqlite3
# import matplotlib.pyplot as plt
import numpy as np
# import ipaddress
# import pygeoip

## Prepara Data Frames Users y Answers

# connection = sqlite3.connect('site.db')
# cursor = connection.cursor()

# query = cursor.execute('''
# SELECT *
# FROM Users 
# ''')

# connection.close()

# headers = [column[0] for column in query.description]

# users_df = pd.DataFrame(query.fetchall(), columns=headers)
# users_df

# import geoip2.database
# reader = geoip2.database.Reader('GeoLite2-City.mmdb')

# def get_continent(ip_str):
# #     print(ip_str)
#     try:
#         response = reader.city(ip_str)
#     except:
#         continent = None
#     else:
#         continent = response.continent.name
#     return continent

# def get_country(ip_str):
# #     print(ip_str)
#     try:
#         response = reader.city(ip_str)
#     except:
#         country = None
#     else:
#         country = response.country.name
#     return country

# def get_city(ip_str):
# #     print(ip_str)
#     try:
#         response = reader.city(ip_str)
#     except:
#         city = None
#     else:
#         city = response.subdivisions.most_specific.name
#     return city

# users_df["ipbased_continent"] = users_df["ip_address"].apply(lambda x: get_continent(x))
# users_df["ipbased_country"] = users_df["ip_address"].apply(lambda x: get_country(x))
# users_df["ipbased_city"] = users_df["ip_address"].apply(lambda x: get_city(x))

# users_df = users_df.replace({'United States':'USA'})

# def check_country(ip_country, stated_country):
# #     print(ip_str)
#     if ip_country == None:
#         return None
#     elif ip_country == stated_country:
#         return 'Valid'
#     else:
#         return 'Invalid'

# users_df["check_country"] = users_df.apply(lambda x: check_country(x.ipbased_country, x.country), axis=1)

# reader.close()
