==> analysis/99 update from heroku.ipynb <==
import sqlite3
import os
import pandas as pd
import numpy as np

## UPDATE FROM HEROKU

os.chdir('./../04 Wekun')
connection = sqlite3.connect('site.db')
cursor = connection.cursor()

cursor.execute("""
SELECT name
FROM sqlite_schema
WHERE 
    type ='table' AND 
    name NOT LIKE 'sqlite_%'
""").fetchall()

users = pd.read_sql('select * from Users',con=connection)
answers = pd.read_sql('select * from Answers',con=connection)
connection.close()

# query = cursor.execute('''
# SELECT  *
# FROM Users
# ''')
# headers = [column[0] for column in query.description]

# new_users_df = pd.DataFrame(query.fetchall(), columns=headers)
# new_users_df

# query = cursor.execute('''
# PRAGMA table_info(Users)
# ''')
# headers = [column[0] for column in query.description]

# query_df = pd.DataFrame(query.fetchall(), columns=headers)
# query_df

# users_dtype = dict(zip(query_df['name'].values,query_df['type'].values))
# users_dtype['create_at'] = 'DATE'
# # users_dtype

# query = cursor.execute('''
# PRAGMA table_info(Answers)
# ''')
# headers = [column[0] for column in query.description]

# query_df = pd.DataFrame(query.fetchall(), columns=headers)
# query_df

# answers_dtype = dict(zip(query_df['name'].values,query_df['type'].values))
# answers_dtype['create_at'] = 'DATE'
# answers_dtype

# connection.close()

## Update From Heroku

import pandas as pd
from sqlalchemy import create_engine

# follows django database settings format, replace with your own settings
DATABASES = {
    'production':{
        'NAME': 'dc6be6usaurgfs',
        'USER': 'wkywflxcyiwwxo',
        'PASSWORD': 'f94a749d823dcf873af8400ea3864fd808fd05c3d3f062dff93568453b736bda',
        'HOST': 'ec2-54-83-17-151.compute-1.amazonaws.com',
        'PORT': 5432,
    },
}

# choose the database to use
db = DATABASES['production']

# construct an engine connection string
engine_string = "postgresql+psycopg2://{user}:{password}@{host}:{port}/{database}".format(
    user = db['USER'],
    password = db['PASSWORD'],
    host = db['HOST'],
    port = db['PORT'],
    database = db['NAME'],
)

# create sqlalchemy engine
engine = create_engine(engine_string)

# read a table from database into pandas dataframe, replace "tablename" with your table name
heroku_users_df = pd.read_sql_table('Users',engine)
heroku_answers_df = pd.read_sql_table('Answers',engine)
mails_df = pd.read_sql_table('Mails',engine)

heroku_answers_df

# users_df = pd.concat([old_users_df,new_users_df,heroku_users_df],ignore_index=True, sort=True)
# users_df = users_df.reset_index()
# users_df['id'] = users_df.index
# users_df = users_df.drop(columns=['index'])

# users_df = users_df[['id', 'user_id', 'device_id', 'create_at', 'age', 'gender', 'education', 'transport', 
#                'nationality', 'ip_address', 'comuna', 'region', 'country',
#                'browser', 'platform', 'psico1', 'psico2', 'psico3', 'psico4', 'psico5',
#                'psico6', 'psico7', 'psico8', 'psico9', 'psico10', 'income']]

# # users_df

# answers_df = pd.concat([old_answers_df,new_answers_df,heroku_answers_df],ignore_index=True)

# # Corregir multiples clicks
# answers_df = answers_df.groupby(['user_id','img_1','img_2','question']).first()
# answers_df = answers_df.sort_values(['id']).reset_index()
# answers_df['id'] = answers_df.index

# answers_df = answers_df[['id','user_id','img_1','img_2','question','choice','create_at']]
# # answers_df

# answers_df.to_csv('answers_XX-09-2019.csv',index=False)
# users_df.to_csv('users_XX-09-2019.csv',index=False)
# mails_df.to_csv('mails_XX-09-2019.csv',index=False)
